docker/services/product-server/requirements.txt
azure-cosmosdb-table==1.0.6
azure-storage-blob==1.4.0
azure-storage-file-share==12.16.0
beautifulsoup4==4.11.1

bidict==0.22.1
biopython==1.81
boto3==1.28.78
cryptography==35.0.0
dask==2022.1.1
dnspython==2.2.1
eventlet==0.33.3
flasgger==0.9.5
flask==1.0.3
flask-jwt-extended==4.3.0
flask-migrate==2.4.0
flask-script==2.0.6
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.4.0
flask-talisman==1.0.0
flask_cors==3.0.8
gevent==24.2.1
gmaps==0.9.0
google-api-core==2.17.0
google-auth==2.27.0
gunicorn==21.0.0
html-sanitizer==2.2.0
httpcore==1.0.2
httpx==0.25.1
ipython==8.21.0
ipywidgets==7.6.3
itsdangerous==2.0.1
Jinja2==3.0.3
json5==0.9.14
jupyterlab==4.1.1
jupyterlab-server==2.25.2
kaleido==0.2.1
kfp==1.8.22
langchain==0.2.3
lxml_html_clean==0.1.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.8.2
millify==0.1.1
mistune==0.8.4
msal==1.26.0
msrest==0.7.1
msrestazure==0.6.4
natsort==8.4.0
networkx==3.1
notebook-shim==0.2.3
numpy==1.24.2
openai==1.3.4
opencensus-ext-azure==1.1.9
opencensus-ext-flask==0.8.2
opencv-python-headless==4.7.0.72
openpyxl==3.0.7
pandas==1.5.3
parquet==1.3.1
plotly==5.8.0
prometheus-client==0.19.0
psutil==5.8.0
# psycopg2==2.8.4
psycopg2-binary==2.9.5
pyarrow==7.0.0
pycountry==22.3.5
pydantic==1.10.8
pyjwt==2.4.0
pymongo==4.3.3
pyodbc==4.0.39
PyPDF2==3.0.1
scikit-learn==1.2.1
scipy == 1.10.1
sendgrid==6.9.7
sentry-sdk[flask]==1.24.0
simple-websocket
snowflake-connector-python==2.7.6
SQLAlchemy==1.4.46
statsmodels == 0.13.5
threadpoolctl==3.2.0
Werkzeug==2.0.3
wordcloud==1.9.2
xlrd==2.0.1
